trading-system/backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.3
redis==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
websockets==11.0.3
httpx==0.27.0
